requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
akinator.py==5.0.0
alt-discord-flags==0.0.5
anyio==3.6.1
appdirs==1.4.4
astunparse==1.6.3
async-8ball==0.1
async-cse==0.3.0
async-generator==1.10
async-timeout==3.0.1
asyncdagpi==4.0.0
attrs==21.2.0
beautifulsoup4==4.10.0
better-profanity==0.7.0
bezier==2021.2.12
blinker==1.5
braceexpand==0.1.7
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
chatbotAI==0.3.1.1
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
discord==1.7.3
discord-ext-flags==0.0.1a0
discord-ext-ipc==2.1.1
discord-ext-menus==1.0.0a31+gfbb8803
discord-ext-ui==2.1.2
discord-py-slash-command==3.0.1
discord.py @ git+https://github.com/Rapptz/discord.py@c26473d0ebcc5f06736d6459b389b2631bf69b7b
docopt==0.6.2
eight-ball==3.0.3
emoji==1.6.1
expr.py==0.3.0
fastapi==0.79.0
Flask==2.1.3
Folders==0.0.1
fonttools==4.34.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httptools==0.4.0
humanize==3.13.1
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.21.0
import-expression==1.1.4
importlib-metadata==4.8.2
itsdangerous==2.1.2
Jinja2==3.1.2
jishaku @ git+https://github.com/Gorialis/jishaku@108f2741e4eb38996584910796d9a8f8bf4c82d7
keyboard==0.13.5
kiwisolver==1.4.4
lyricsgenius==3.0.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mouse==0.7.1
MouseInfo==0.1.3
mpmath==1.2.1
multidict==5.2.0
mystbin.py==2.2.0
nose==1.3.7
numpy==1.23.1
opencv-python==4.6.0.66
opuslib==3.0.3
outcome==1.1.0
packaging==21.3
Panda3D==1.10.10
Pillow==8.4.0
pilmoji==2.0.1
pipreqs==0.4.11
ply==3.11
praw==7.5.0
prawcore==2.3.0
prettify==0.1.1
prettify.py @ git+https://github.com/ShineyDev/prettify.py@8e3440930e08ea309c124ec654eb70bf5ba9734c
pretty-errors==1.2.25
priority==2.0.0
psutil==5.9.1
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.9.1
pygame==2.1.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.20.0
pytio==0.3.1
pytweening==1.0.4
pyvista==0.36.0
PyYAML==6.0
quart==0.17.0
Random-Word==1.0.7
requests==2.26.0
rply==0.7.8
scipy==1.9.0
scooby==0.6.0
selenium==4.1.0
singleton==0.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.19.1
sympy==1.10.1
tabulate==0.8.10
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing==3.7.4.3
typing_extensions==4.0.1
udpy==2.0.0
update-checker==0.18.0
upsidedown==0.4
urllib3==1.26.7
uvicorn==0.18.2
vtk==9.1.0
Wand==0.6.8
watchfiles==0.15.0
websocket-client==1.2.3
websockets==10.3
Werkzeug==2.1.2
wikipedia==1.4.0
wslink==1.6.6
wsproto==1.0.0
yarg==0.1.9
yarl==1.7.2
youtube-dl==2021.6.6
zipp==3.6.0
